+++ src/pages/test/main.ts
@@ -4,11 +4,15 @@ import {config} from "../../../plugin";
 
 import country from "./widget-input-country.vue";
 import accountType from "@/pages/test/widget-input-account-type.vue";
+import InputCoord from "@/pages/test/input-coord.vue";
+import InputFile from "@/pages/test/input-file.vue";
 
 config({
 	inputTypes: {
 		country,
-		"account-type": accountType
+		"account-type": accountType,
+		coord: InputCoord,
+		file: InputFile
 	},
 	debug: true
 })